Add accumulate function and tests
Implement a function that performs a left fold operation on an iterator
with a given initial value and a closure. This function is similar to
std::iter::Iterator::fold, but it takes the closure as the last argument
and allows using the function name as a higher-order function.

Write unit tests for the accumulate function using different operations,
such as sum, product, and GCD. Also test an empty iterator case with
#[should_panic] attribute.

This function can be useful for applying various aggregations on
collections of values in a concise and functional way.